refactor(eve): split task workflows into parent and child modules
