+++ packages/eve/src/execution/dispatch-runtime-actions-step.ts
@@ -1,271 +1,71 @@
 /**
- * Dispatches every pending runtime action for a parked parent session.
- * Starts commit ownership before their side effect, so the returned state
- * owns every child it may have created.
+ * Starts or continues every pending runtime action for the parked parent
+ * session, with children reporting straight back to the parent turn via
+ * `parentContinuationToken`.
+ *
+ * The batch is classified into a dispatch plan first (reject / resume /
+ * start), then each entry dispatches and emits one
+ * parent `subagent.called` control-plane event through a single tail.
+ * Every start commits an agent handle (`starting`) before its side effect
+ * and confirms it (`running`) once the child reports coordinates, so the
+ * returned snapshot-bearing state owns every child it may have created.
+ *
+ * Agents running `experimental.tasks` never reach this step: the turn
+ * workflow selects `dispatchTaskStep`, the task-mode sibling that wraps
+ * each dispatch in the delegated-task lifecycle.
  */
 
-import { buildAdapterContext } from "#channel/adapter-context.js";
-import { callAdapterEventHandler } from "#channel/adapter.js";
-import {
-  AuthKey,
-  CapabilitiesKey,
-  ChannelInstrumentationKey,
-  InitiatorAuthKey,
-} from "#context/keys.js";
-import {
-  BundleKey,
-  ChannelKey,
-  type CompiledBundle,
-} from "#runtime/sessions/runtime-context-keys.js";
-import { deserializeContext } from "#context/serialize.js";
-import {
-  dispatchToAgentHandle,
-  dispatchToTaskAgentAddress,
-  isAgentHandleAction,
-  type DispatchOutcome,
-  type RuntimeAgentHandleAction,
-  type RuntimeSession,
-} from "#execution/agent-handle-dispatch.js";
+import { type DispatchOutcome, dispatchToAgentHandle } from "#execution/agent-handle-dispatch.js";
 import { createAgentContinuationBundle } from "#execution/agent-continuation-bundle.js";
-import { getAgentHandleStore } from "#harness/handles/store.js";
 import {
-  assertUniqueRuntimeActionCallIds,
-  getPendingRuntimeActionBatch,
-} from "#harness/runtime-actions.js";
-import {
-  createSubagentCalledEvent,
-  encodeMessageStreamEvent,
-  stampMessageStreamEvent,
-} from "#protocol/message.js";
-import type {
-  RuntimeActionRequest,
-  RuntimeActionResult,
-  RuntimeRemoteAgentCallActionRequest,
-  RuntimeSubagentCallActionRequest,
-  RuntimeSubagentDispatchFailure,
-  RuntimeToolCallActionRequest,
-} from "#runtime/actions/types.js";
-import {
-  beginDelegatedTask,
-  type DelegatedTask,
-  executeTaskControlAction,
-  isTaskControlAction,
-  settleDelegatedDispatch,
-} from "#execution/tasks/dispatch.js";
-import {
-  createDurableSessionState,
-  type DurableSessionState,
-  readDurableSession,
-} from "#execution/durable-session-store.js";
-import {
-  createRecursiveAgentRootOnlyResult,
-  createUnavailableDynamicSubagentResult,
-  getSubagentName,
-} from "#execution/dispatch-action-failures.js";
-import { startLocalSubagent } from "#execution/subagent-start-local.js";
-import { startRemoteSubagent } from "#execution/subagent-start-remote.js";
-import { hydrateDurableSession } from "#execution/session.js";
-import { buildSubagentRunInput, type SubagentInputSource } from "#execution/subagent-tool.js";
-import { workflowEntryReference } from "#execution/workflow-runtime.js";
-import { createLogger, logError } from "#internal/logging.js";
-import { readSessionTraceContext } from "#tracing/agent-trace-context-store.js";
-import { resolveSubagentDepth } from "#harness/subagent-depth.js";
-import { getDynamicSubagentSelection } from "#context/dynamic-subagent-lifecycle.js";
-import { resolveEffectiveAgentRuntime } from "#execution/effective-agent-config.js";
-import {
-  checkTaskContinuationAdmission,
-  describeTaskDispatch,
-  reserveTaskContinuation,
-  settleTaskDispatchError,
-  type ReservedTaskContinuation,
-} from "#execution/tasks/continuation-admission.js";
+  emitSubagentCalled,
+  prepareRuntimeActionDispatch,
+  type RuntimeActionDispatchInput,
+  type RuntimeActionDispatchResult,
+  startSubagent,
+} from "#execution/dispatch-runtime-actions-shared.js";
+import { createDurableSessionState } from "#execution/durable-session-store.js";
+import type { RuntimeActionResult } from "#runtime/actions/types.js";
 
-const log = createLogger("execution.dispatch-runtime-actions");
-
-type DynamicSubagentAgentConfig = NonNullable<
-  Extract<
-    ReturnType<typeof getDynamicSubagentSelection>,
-    { readonly kind: "subagent" }
-  >["agentConfig"]
->;
-
-type DynamicRemoteAgentConfig = NonNullable<
-  Extract<
-    ReturnType<typeof getDynamicSubagentSelection>,
-    { readonly kind: "remote" }
-  >["remoteAgent"]
->;
-
-type DispatchPlanEntry =
-  | {
-      readonly kind: "resume";
-      readonly action: RuntimeAgentHandleAction;
-      readonly agentId: string;
-      readonly dynamicRemoteAgent?: DynamicRemoteAgentConfig;
-    }
-  | { readonly kind: "reject"; readonly result: RuntimeSubagentDispatchFailure }
-  | { readonly kind: "start"; readonly target: DispatchStartTarget }
-  | { readonly kind: "task-control"; readonly action: RuntimeToolCallActionRequest };
-
-type DispatchStartTarget =
-  | {
-      readonly kind: "local";
-      readonly action: RuntimeSubagentCallActionRequest;
-      readonly dynamicSubagentAgentConfig?: DynamicSubagentAgentConfig;
-      readonly source: SubagentInputSource;
-    }
-  | {
-      readonly kind: "remote";
-      readonly action: RuntimeRemoteAgentCallActionRequest;
-      readonly dynamicRemoteAgent?: DynamicRemoteAgentConfig;
-    };
-
-export async function dispatchRuntimeActionsStep(input: {
-  readonly callbackBaseUrl?: string;
-  /** Internal hook that receives child completion and HITL payloads. */
-  readonly parentContinuationToken?: string;
-  readonly parentWritable: WritableStream<Uint8Array>;
-  readonly serializedContext: Record<string, unknown>;
-  readonly sessionState: DurableSessionState;
-}): Promise<{
-  readonly results: readonly RuntimeActionResult[];
-  readonly sessionState: DurableSessionState;
-  readonly taskReadiness: readonly {
-    readonly continuationToken: string;
-    readonly taskId: string;
-    readonly taskRunId: string;
-  }[];
-}> {
+export async function dispatchRuntimeActionsStep(
+  input: RuntimeActionDispatchInput,
+): Promise<RuntimeActionDispatchResult> {
   "use step";
 
-  const durableSession = await readDurableSession(input.sessionState);
-  const batch = getPendingRuntimeActionBatch(durableSession.state);
-
-  if (batch === undefined || batch.actions.length === 0) {
+  const prepared = await prepareRuntimeActionDispatch({
+    serializedContext: input.serializedContext,
+    sessionState: input.sessionState,
+    taskControls: false,
+  });
+  if (prepared === undefined) {
     return { results: [], sessionState: input.sessionState, taskReadiness: [] };
   }
-  assertUniqueRuntimeActionCallIds(batch.actions);
 
-  const parentTurnId = batch.event.turnId || activeTurnId(input.sessionState.emissionState);
-  const batchEvent = { ...batch.event, turnId: parentTurnId };
-
-  const ctx = await deserializeContext(input.serializedContext);
-  const bundle = ctx.require(BundleKey);
-  const effectiveAgent = resolveEffectiveAgentRuntime(bundle, ctx);
-  const session = setPendingRuntimeActionBatch({
-    ...batch,
-    event: batchEvent,
-    session: hydrateDurableSession({
-      compactionOverrides: { thresholdPercent: effectiveAgent.thresholdPercent },
-      durable: durableSession,
-      turnAgent: effectiveAgent.turnAgent,
-    }),
-  });
-  const adapter = ctx.require(ChannelKey);
-  const auth = ctx.get(AuthKey) ?? null;
-  const capabilities = ctx.get(CapabilitiesKey);
-  const channelMetadata = ctx.get(ChannelInstrumentationKey);
-  const initiatorAuth = ctx.get(InitiatorAuthKey) ?? null;
-
-  const adapterCtx = buildAdapterContext(adapter, ctx);
-  const parentTraceContext = readSessionTraceContext(input.serializedContext, session.sessionId);
-  const tasksEnabled = bundle.resolvedAgent.config.experimental?.tasks === true;
-  // Background tasks require resumable children: the flag implies
-  // conversation-mode dispatch so `experimental.tasks` and
-  // `experimental.subagentPersistentSessions` never produce a third mode.
+  const { batch, bundle, session } = prepared;
   const persistentSessions =
-    tasksEnabled || bundle.resolvedAgent.config.experimental?.subagentPersistentSessions === true;
-  // A corrupt handle store throws; surface that before anything dispatches.
-  // A mid-loop throw after a sibling started would durably replay the whole
-  // batch and re-dispatch that sibling.
-  getAgentHandleStore(durableSession.state);
-  const plan = planDispatch({ actions: batch.actions, bundle, ctx, session });
+    bundle.resolvedAgent.config.experimental?.subagentPersistentSessions === true;
   // Acquired only once preflight can no longer throw, so a planning failure
   // never leaks the writer lock.
   const writer = input.parentWritable.getWriter();
-  // Split the parent's remaining token quota across the batch's freshly
-  // started local subagents, the children that actually receive an enforced
-  // cap. Continuations already run under their own budget, and remote agents
-  // run on their own deployment under their own limits, so neither dilutes
-  // the local shares.
-  const fanoutSize = plan.filter(
-    (entry) => entry.kind === "start" && entry.target.kind === "local",
-  ).length;
 
   let nextSession = session;
   const results: RuntimeActionResult[] = [];
-  const taskReadiness: DelegatedTask[] = [];
 
   try {
-    for (const entry of plan) {
+    for (const entry of prepared.plan) {
       if (entry.kind === "reject") {
         results.push(entry.result);
         continue;
       }
-
       if (entry.kind === "task-control") {
-        const control = await executeTaskControlAction({
-          action: entry.action,
-          bundle,
-          parentStepIndex: batch.event.stepIndex,
-          parentTurnId: batch.event.turnId,
-          session: nextSession,
-        });
-        nextSession = control.session;
-        if (control.taskReadiness !== undefined) taskReadiness.push(control.taskReadiness);
-        if (control.result !== undefined) {
-          results.push(control.result);
-        }
-        continue;
-      }
-
-      if (tasksEnabled && entry.kind === "resume") {
-        const busy = await checkTaskContinuationAdmission({
-          action: entry.action,
-          agentId: entry.agentId,
-          parentStepIndex: batch.event.stepIndex,
-          parentTurnId: batch.event.turnId,
-          session: nextSession,
-        });
-        if (busy !== undefined) {
-          results.push(busy);
-          continue;
-        }
-      }
-
-      const delegatedAction = entry.kind === "resume" ? entry.action : entry.target.action;
-      const delegatedDescription = describeTaskDispatch({
-        action: delegatedAction,
-        agentId: entry.kind === "resume" ? entry.agentId : undefined,
-        parentSessionId: session.sessionId,
-        parentTurnId: batch.event.turnId,
-        session: nextSession,
-      });
-      const delegated = tasksEnabled
-        ? await beginDelegatedTask({
-            ...delegatedDescription,
-            parentSessionId: session.sessionId,
-            parentStepIndex: batch.event.stepIndex,
-            parentTurnId: batch.event.turnId,
-            session: nextSession,
-          })
-        : undefined;
-      const delegatedParentToken = delegated?.continuationToken;
-      let reservedContinuation: ReservedTaskContinuation | undefined;
-      if (entry.kind === "resume") {
-        reservedContinuation = await reserveTaskContinuation({
-          action: entry.action,
-          agentId: entry.agentId,
-          delegated,
-          session: nextSession,
-        });
-        nextSession = reservedContinuation?.session ?? nextSession;
+        // Unreachable: the plan was built with `taskControls: false`.
+        throw new Error("Task-control actions require the task dispatch step.");
       }
 
       let outcome: DispatchOutcome;
       switch (entry.kind) {
         case "resume":
-          outcome = await (tasksEnabled ? dispatchToTaskAgentAddress : dispatchToAgentHandle)({
+          outcome = await dispatchToAgentHandle({
             action: entry.action,
             agentId: entry.agentId,
             bundle: createAgentContinuationBundle({
@@ -274,27 +74,26 @@ export async function dispatchRuntimeActionsStep(input: {
               dynamicRemoteAgent: entry.dynamicRemoteAgent,
             }),
             currentSession: nextSession,
-            parentToken:
-              delegatedParentToken ?? input.parentContinuationToken ?? session.continuationToken,
+            parentToken: input.parentContinuationToken ?? session.continuationToken,
             parentTurnId: batch.event.turnId,
           });
           break;
         case "start":
           outcome = await startSubagent({
-            auth,
-            batchEvent,
+            auth: prepared.auth,
+            batchEvent: batch.event,
             bundle,
             callbackBaseUrl: input.callbackBaseUrl,
-            capabilities,
-            channelMetadata,
+            capabilities: prepared.capabilities,
+            channelMetadata: prepared.channelMetadata,
             currentSession: nextSession,
-            fanoutSize,
-            initiatorAuth,
-            parentContinuationToken: delegatedParentToken ?? input.parentContinuationToken,
-            parentTraceContext,
+            fanoutSize: prepared.fanoutSize,
+            initiatorAuth: prepared.initiatorAuth,
+            parentContinuationToken: input.parentContinuationToken,
+            parentTraceContext: prepared.parentTraceContext,
             persistentSessions,
             session,
-            taskOwned: tasksEnabled,
+            taskOwned: false,
             target: entry.target,
           });
           break;
@@ -302,314 +101,30 @@ export async function dispatchRuntimeActionsStep(input: {
 
       nextSession = outcome.session;
       if (outcome.kind === "error") {
-        results.push(
-          await settleTaskDispatchError({
-            agentId: entry.kind === "resume" ? entry.agentId : undefined,
-            delegated,
-            outcome,
-            reserved: reservedContinuation,
-            session: nextSession,
-          }),
-        );
-        if (delegated !== undefined) taskReadiness.push(delegated);
+        results.push(outcome.result);
         continue;
       }
 
-      if (delegated !== undefined) {
-        if (reservedContinuation !== undefined) {
-          results.push(reservedContinuation.receipt);
-        } else {
-          const settled = await settleDelegatedDispatch({
-            callId: outcome.callId,
-            session: nextSession,
-            subagentName: outcome.toolName,
-            task: delegated,
-          });
-          nextSession = settled.session;
-          results.push(settled.receipt);
-        }
-        taskReadiness.push(delegated);
-      }
-
-      // Emission is observability, not control flow: a failure here must not
-      // escape the loop, because a durable-step retry would re-dispatch the
-      // children that already started.
-      try {
-        const action = entry.kind === "resume" ? entry.action : entry.target.action;
-        const dynamicRemoteAgent =
-          entry.kind === "resume"
-            ? entry.dynamicRemoteAgent
-            : entry.target.kind === "remote"
-              ? entry.target.dynamicRemoteAgent
-              : undefined;
-        const parentEvent = await callAdapterEventHandler(
-          adapter,
-          createSubagentCalledEvent({
-            callId: outcome.callId,
-            childSessionId: outcome.address.sessionId,
-            name: outcome.name,
-            remote:
-              outcome.address.kind === "agent/remote"
-                ? {
-                    resolverId:
-                      dynamicRemoteAgent === undefined
-                        ? action.nodeId
-                        : dynamicRemoteAgent.credentialsStepId,
-                    url: outcome.address.url,
-                  }
-                : undefined,
-            sequence: batch.event.sequence,
-            sessionId: session.sessionId,
-            toolName: outcome.toolName,
-            turnId: parentTurnId,
-            workflowId: workflowEntryReference.workflowId,
-          }),
-          adapterCtx,
-        );
-        await writer.write(encodeMessageStreamEvent(stampMessageStreamEvent(parentEvent)));
-      } catch (error) {
-        logError(log, "subagent.called emission failed", error, {
-          callId: outcome.callId,
-          childSessionId: outcome.address.sessionId,
-          toolName: outcome.toolName,
-        });
-      }
+      await emitSubagentCalled({
+        adapter: prepared.adapter,
+        adapterCtx: prepared.adapterCtx,
+        batchEvent: batch.event,
+        entry,
+        outcome,
+        sessionId: session.sessionId,
+        writer,
+      });
     }
   } finally {
     writer.releaseLock();
   }
 
-  const nextState =
-    nextSession === session && batch.event.turnId === parentTurnId
-      ? input.sessionState
-      : createDurableSessionState({ session: nextSession });
-
   return {
     results,
-    sessionState: nextState,
-    taskReadiness,
+    sessionState:
+      nextSession === session
+        ? input.sessionState
+        : createDurableSessionState({ session: nextSession }),
+    taskReadiness: [],
   };
-}
-
-/**
- * Classifies every batch action before anything dispatches, so invalid
- * batches fail without starting children and rejections never interleave
- * with dispatch work.
- *
- * This is the single place that decides fresh start vs. continuation:
- * an omitted, null, empty, or whitespace-only agentId is a fresh start
- * (strict tool-calling providers force every schema property to be present,
- * so models emit `""`/`null` when they mean "no continuation"), and an
- * agentId that matches no stored handle also falls back to a fresh start —
- * models sometimes pass a hallucinated or stale id, and hard-failing made
- * them conclude the subagent itself was unavailable. Only an id that
- * resolves to a stored handle becomes a resume.
- */
-function planDispatch(input: {
-  readonly actions: readonly RuntimeActionRequest[];
-  readonly bundle: CompiledBundle;
-  readonly ctx: Parameters<typeof getDynamicSubagentSelection>[0];
-  readonly session: RuntimeSession;
-}): DispatchPlanEntry[] {
-  const handles = getAgentHandleStore(input.session.state)?.handles ?? [];
-
-  return input.actions.map((action): DispatchPlanEntry => {
-    if (isTaskControlAction(action)) {
-      return { action, kind: "task-control" };
-    }
-
-    const rawAgentId = action.input.agentId;
-    const agentId =
-      typeof rawAgentId === "string" && rawAgentId.trim() !== "" ? rawAgentId : undefined;
-    if (agentId !== undefined && isAgentHandleAction(action)) {
-      // Resume classification runs before the recursion guard: an agentId
-      // continuation resumes an already-adopted child rather than starting
-      // a new one. Unknown ids go through classifyFreshStart below, which
-      // re-applies the guard the resume path bypasses.
-      if (handles.some((handle) => handle.identity.id === agentId)) {
-        const dynamicSubagentSelection =
-          input.bundle.subagentRegistry.dynamicNodeIds?.has(action.nodeId) === true
-            ? getDynamicSubagentSelection(input.ctx, action.nodeId)
-            : undefined;
-        return {
-          action,
-          agentId,
-          dynamicRemoteAgent:
-            action.kind === "remote-agent-call" && dynamicSubagentSelection?.kind === "remote"
-              ? dynamicSubagentSelection.remoteAgent
-              : undefined,
-          kind: "resume",
-        };
-      }
-      log.warn("unknown agentId on subagent call; starting a new agent", {
-        agentId,
-        callId: action.callId,
-      });
-    }
-
-    return classifyFreshStart({
-      action,
-      bundle: input.bundle,
-      ctx: input.ctx,
-      session: input.session,
-    });
-  });
-}
-
-/**
- * Classifies one action for fresh dispatch: rejected by the recursion guard,
- * or started against a local/remote target. Shared by plain starts and the
- * unknown-agentId fallback, so both paths enforce the same guard.
- */
-function classifyFreshStart(input: {
-  readonly action: RuntimeActionRequest;
-  readonly bundle: CompiledBundle;
-  readonly ctx: Parameters<typeof getDynamicSubagentSelection>[0];
-  readonly session: RuntimeSession;
-}): Extract<DispatchPlanEntry, { kind: "reject" | "start" }> {
-  const { action } = input;
-  const registry = input.bundle.subagentRegistry.subagentsByNodeId;
-  const subagentDepth = resolveSubagentDepth(input.session);
-  const rootOnly = input.session.rootSessionId !== undefined || subagentDepth.currentDepth > 0;
-
-  const isDynamicSubagent =
-    (action.kind === "subagent-call" || action.kind === "remote-agent-call") &&
-    input.bundle.subagentRegistry.dynamicNodeIds?.has(action.nodeId) === true;
-  const dynamicSubagentSelection = isDynamicSubagent
-    ? getDynamicSubagentSelection(input.ctx, action.nodeId)
-    : undefined;
-  if (
-    isDynamicSubagent &&
-    (dynamicSubagentSelection === undefined ||
-      (action.kind === "subagent-call" && dynamicSubagentSelection.kind !== "subagent") ||
-      (action.kind === "remote-agent-call" && dynamicSubagentSelection.kind !== "remote"))
-  ) {
-    const subagentName = getSubagentName(action);
-    log.warn("dynamic subagent call blocked after availability changed", {
-      callId: action.callId,
-      nodeId: action.nodeId,
-      subagentName,
-    });
-    return { kind: "reject", result: createUnavailableDynamicSubagentResult(action) };
-  }
-
-  if (isRecursiveAgentAction(action, registry) && rootOnly) {
-    log.warn("recursive agent call blocked outside the root session", {
-      callId: action.callId,
-      currentDepth: subagentDepth.currentDepth,
-      nodeId: action.nodeId,
-      subagentName: action.subagentName,
-    });
-    return { kind: "reject", result: createRecursiveAgentRootOnlyResult(action) };
-  }
-
-  switch (action.kind) {
-    case "subagent-call": {
-      const dynamicAgentConfig =
-        dynamicSubagentSelection?.kind === "subagent"
-          ? dynamicSubagentSelection.agentConfig
-          : undefined;
-      const registered = registry.get(action.nodeId);
-      const description =
-        dynamicAgentConfig?.description ??
-        (registered?.definition.kind === "subagent"
-          ? registered.definition.description
-          : undefined);
-      const source: SubagentInputSource =
-        description !== undefined ? { description, type: "local" } : { type: "runtime" };
-      return {
-        kind: "start",
-        target: {
-          action,
-          dynamicSubagentAgentConfig: dynamicAgentConfig,
-          kind: "local",
-          source,
-        },
-      };
-    }
-    case "remote-agent-call":
-      return {
-        kind: "start",
-        target: {
-          action,
-          dynamicRemoteAgent:
-            dynamicSubagentSelection?.kind === "remote"
-              ? dynamicSubagentSelection.remoteAgent
-              : undefined,
-          kind: "remote",
-        },
-      };
-    default:
-      throw new Error(`Unsupported runtime action kind "${action.kind}" in workflow runtime.`);
-  }
-}
-
-async function startSubagent(input: {
-  readonly auth: Parameters<typeof buildSubagentRunInput>[0]["auth"];
-  readonly batchEvent: { readonly sequence: number; readonly turnId: string };
-  readonly bundle: CompiledBundle;
-  readonly callbackBaseUrl: string | undefined;
-  readonly capabilities: Parameters<typeof buildSubagentRunInput>[0]["capabilities"];
-  readonly channelMetadata: Parameters<typeof buildSubagentRunInput>[0]["channelMetadata"];
-  readonly currentSession: RuntimeSession;
-  readonly fanoutSize: number;
-  readonly initiatorAuth: Parameters<typeof buildSubagentRunInput>[0]["initiatorAuth"];
-  readonly parentContinuationToken: string | undefined;
-  readonly parentTraceContext: Parameters<typeof buildSubagentRunInput>[0]["parentTraceContext"];
-  readonly persistentSessions: boolean;
-  readonly session: RuntimeSession;
-  readonly taskOwned: boolean;
-  readonly target: DispatchStartTarget;
-}): Promise<DispatchOutcome> {
-  switch (input.target.kind) {
-    case "local":
-      return startLocalSubagent({
-        action: input.target.action,
-        auth: input.auth,
-        batchEvent: input.batchEvent,
-        bundle: input.bundle,
-        capabilities: input.capabilities,
-        channelMetadata: input.channelMetadata,
-        currentSession: input.currentSession,
-        dynamicSubagentAgentConfig: input.target.dynamicSubagentAgentConfig,
-        fanoutSize: input.fanoutSize,
-        initiatorAuth: input.initiatorAuth,
-        parentContinuationToken: input.parentContinuationToken,
-        parentTraceContext: input.parentTraceContext,
-        persistentSessions: input.persistentSessions,
-        session: input.session,
-        source: input.target.source,
-        taskOwned: input.taskOwned,
-      });
-    case "remote":
-      return startRemoteSubagent({
-        action: input.target.action,
-        auth: input.auth,
-        batchEvent: input.batchEvent,
-        bundle: input.bundle,
-        callbackBaseUrl: input.callbackBaseUrl,
-        currentSession: input.currentSession,
-        dynamicRemoteAgent: input.target.dynamicRemoteAgent,
-        initiatorAuth: input.initiatorAuth,
-        parentContinuationToken: input.parentContinuationToken,
-        persistentSessions: input.persistentSessions,
-        session: input.session,
-        taskOwned: input.taskOwned,
-      });
-    default: {
-      const _exhaustive: never = input.target;
-      return _exhaustive;
-    }
-  }
-}
-
-function isRecursiveAgentAction(
-  action: RuntimeActionRequest,
-  subagentsByNodeId: ReadonlyMap<string, unknown>,
-): action is RuntimeSubagentCallActionRequest {
-  return (
-    action.kind === "subagent-call" &&
-    action.subagentName === "agent" &&
-    !subagentsByNodeId.has(action.nodeId)
-  );
 }
