+++ packages/eve/src/execution/turn-execution-cursor.ts
@@ -6,6 +6,7 @@ import type {
   TurnStepInput,
   TurnStepPayload,
 } from "#execution/durable-session-migrations/turn-workflow.js";
+import { SessionStateCursor } from "#execution/session-state-cursor.js";
 import type { SettledTurn } from "#harness/types.js";
 import type { TokenUsage } from "#shared/token-usage.js";
 
@@ -31,12 +32,10 @@ type TurnTerminalAction =
     };
 
 /** Owns the mutable durable state cursor for one active turn workflow. */
-export class TurnExecutionCursor {
+export class TurnExecutionCursor extends SessionStateCursor {
   readonly controlToken: string;
   readonly parentWritable: WritableStream<Uint8Array>;
 
-  private currentSerializedContext: Record<string, unknown>;
-  private currentSessionState: DurableSessionState;
   private lastReportedContinuationToken: string;
 
   constructor(input: {
@@ -45,26 +44,15 @@ export class TurnExecutionCursor {
     readonly serializedContext: Record<string, unknown>;
     readonly sessionState: DurableSessionState;
   }) {
+    super({ serializedContext: input.serializedContext, sessionState: input.sessionState });
     this.controlToken = input.controlToken;
-    this.currentSerializedContext = input.serializedContext;
-    this.currentSessionState = input.sessionState;
     this.lastReportedContinuationToken = input.sessionState.continuationToken;
     this.parentWritable = input.parentWritable;
   }
 
-  /** Latest serialized runtime context adopted by the active turn. */
-  get serializedContext(): Record<string, unknown> {
-    return this.currentSerializedContext;
-  }
-
-  /** Latest durable session state adopted by the active turn. */
-  get sessionState(): DurableSessionState {
-    return this.currentSessionState;
-  }
-
   /** Adopts a state transition and reports any continuation-token change once. */
   async adopt(transition: TurnTransition): Promise<void> {
-    this.setState(transition);
+    this.adoptState(transition);
 
     const nextToken = transition.sessionState.continuationToken;
     if (nextToken === "" || nextToken === this.lastReportedContinuationToken) return;
@@ -79,8 +67,8 @@ export class TurnExecutionCursor {
       abortSignal,
       input,
       parentWritable: this.parentWritable,
-      serializedContext: this.currentSerializedContext,
-      sessionState: this.currentSessionState,
+      serializedContext: this.serializedContext,
+      sessionState: this.sessionState,
     };
   }
 
@@ -94,12 +82,12 @@ export class TurnExecutionCursor {
     action: TurnTerminalAction,
     bufferedDeliveries: readonly DeliverHookPayload[],
   ): Promise<void> {
-    this.setState(transition);
+    this.adoptState(transition);
     await this.send({
       action: {
         ...action,
-        serializedContext: this.currentSerializedContext,
-        sessionState: this.currentSessionState,
+        serializedContext: this.serializedContext,
+        sessionState: this.sessionState,
       },
       bufferedDeliveries: bufferedDeliveries.length === 0 ? undefined : [...bufferedDeliveries],
       kind: "turn-result",
@@ -109,10 +97,5 @@ export class TurnExecutionCursor {
   /** Sends one control payload to the session driver. */
   async send(payload: TurnControlPayload): Promise<void> {
     await sendTurnControlStep({ controlToken: this.controlToken, payload });
-  }
-
-  private setState(transition: TurnTransition): void {
-    this.currentSerializedContext = transition.serializedContext ?? this.currentSerializedContext;
-    this.currentSessionState = transition.sessionState;
   }
 }
